src/test/java/_0801_0850/_813_Largest_Sum_of_Averages/SolutionITest.java
package _0801_0850._813_Largest_Sum_of_Averages;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 8/31/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int A[] = {9, 1, 2, 3, 9}, K = 3;
    assertEquals(20, solution.largestSumOfAverages(A, K));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}